Unify JAR location in all openvidu-server Dockerfiles

+++ openvidu-server/docker/openvidu-dev/create_image.sh
@@ -1,7 +1,15 @@
 #!/bin/bash
 VERSION=$1
 if [[ ! -z $VERSION ]]; then
-    cp ../../target/openvidu-server-*.jar ./openvidu-server.jar
+
+    if [[ ! -f "./openvidu-server-*.jar" ]]; then
+        cp ../../target/openvidu-server-*.jar .
+    fi
+    if [[ ! -f "./openvidu-server-*.jar" ]]; then
+        echo "Error: openvidu-server JAR not found"
+        exit 1
+    fi
+
     docker build --pull --no-cache --rm=true -t openvidu/openvidu-dev:$VERSION .
     rm ./openvidu-server.jar
 else
