+++ openvidu-server/docker/openvidu-server/create_image.sh
@@ -1,7 +1,15 @@
 #!/bin/bash
 VERSION=$1
 if [[ ! -z $VERSION ]]; then
-    cp ../../target/openvidu-server-*.jar ./openvidu-server.jar
+
+    if [[ ! -f "./openvidu-server-*.jar" ]]; then
+        cp ../../target/openvidu-server-*.jar .
+    fi
+    if [[ ! -f "./openvidu-server-*.jar" ]]; then
+        echo "Error: openvidu-server JAR not found"
+        exit 1
+    fi
+
     cp ../utils/discover_my_public_ip.sh ./discover_my_public_ip.sh
     cp ../utils/coturn-shared-key.template ./coturn-shared-key.template
